Move preview IDs to a separate file
