+++ app/src/docs/Tag.doc.tsx
@@ -14,13 +14,7 @@ import {
 } from '@epam/uui-docs';
 import { EditableDocContent, DocExample, BaseDocsBlock } from '../common';
 import { getCurrentTheme } from '../helpers';
-
-enum TTagPreview {
-    'One-line caption' = 'One-line caption',
-    'Two-line caption' = 'Two-line caption',
-    'No caption' = 'No caption',
-    Colors = 'Colors'
-}
+import { TTagPreview } from './_types/previewIds';
 
 export class TagDoc extends BaseDocsBlock {
     title = 'Tag';
