+++ uui-e2e-tests/src/data/testData.ts
@@ -1,3 +1,12 @@
+import {
+    TAccordionPreview, TAlertPreview,
+    TAvatarStackPreview,
+    TBadgePreview,
+    TButtonPreview, TCheckboxPreview, TCountIndicatorPreview, TIconButtonPreview,
+    TLinkButtonPreview, TSwitchPreview, TTagPreview,
+    TTextInputPreview,
+} from './previewIds';
+
 /**
  * Keep in sync with app/src/documents/structureComponents.ts
  */
@@ -72,70 +81,6 @@ export enum TComponentId {
     useTableState= 'useTableState',
     verticalTabButton= 'verticalTabButton',
     virtualList= 'virtualList'
-}
-
-/**
- * Copy these enums from the respective *.doc.tsx files
- */
-
-export enum TBadgePreview {
-    'Colors' = 'Colors',
-    'Sizes with icon' = 'Sizes with icon',
-    'Sizes without icon' = 'Sizes without icon'
-}
-export enum TButtonPreview {
-    'One-line caption' = 'One-line caption',
-    'Two-line caption' = 'Two-line caption',
-    'No caption' = 'No caption',
-    Colors = 'Colors'
-}
-export enum TLinkButtonPreview {
-    'One-line caption' = 'One-line caption',
-    'Two-line caption' = 'Two-line caption',
-    'No caption' = 'No caption',
-    Colors = 'Colors'
-}
-export enum TAvatarStackPreview {
-    'Smaller size' = 'Smaller size',
-    'Bigger size' = 'Bigger size'
-}
-export enum TTagPreview {
-    'One-line caption' = 'One-line caption',
-    'Two-line caption' = 'Two-line caption',
-    'No caption' = 'No caption',
-    Colors = 'Colors'
-}
-export enum TSwitchPreview {
-    Basic = 'Basic'
-}
-export enum TCheckboxPreview {
-    Basic = 'Basic'
-}
-export enum TTextInputPreview {
-    'Form'= 'Form',
-    'Form (invalid)'= 'Form (invalid)',
-    'Form (disabled)'= 'Form (disabled)',
-    'Form (read only)'= 'Form (read only)',
-    'Inline'= 'Inline',
-    'Inline (disabled)'= 'Inline (disabled)',
-    'Inline (read only)'= 'Inline (read only)'
-}
-export enum TCountIndicatorPreview {
-    Colors = 'Colors',
-    Sizes = 'Sizes'
-}
-export enum TAccordionPreview {
-    Expanded = 'Expanded',
-    Collapsed = 'Collapsed'
-}
-export enum TAlertPreview {
-    Colors = 'Colors',
-    'Layout with icon' = 'Layout with icon',
-    'Layout without icon' = 'Layout without icon'
-}
-export enum TIconButtonPreview {
-    Colors = 'Colors',
-    Layout = 'Layout'
 }
 
 /**
